fix: consolidate map constants
