--- app/site-config/map/stac-endpoints.ts
@@ -1,4 +1,0 @@
-export const STAC_ENDPOINTS = {
-  stacApiUrl: "https://dev.disasters.openveda.cloud/api/stac",
-  titilerBaseUrl: "https://dev.disasters.openveda.cloud/api/raster",
-};
